+++ historic/pom.xml
@@ -9,7 +9,7 @@
 		<version>2.4.4</version>
 		<relativePath /> <!-- lookup parent from repository -->
 	</parent>
-	<groupId>web.historic</groupId>
+	<groupId>web.mediscreen</groupId>
 	<artifactId>historic</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
 	<name>historic</name>
