+++ historic/src/main/java/web/mediscreen/historic/configuration/MongoDbConfig.java
@@ -1,0 +1,10 @@
+package web.mediscreen.historic.configuration;
+
+import org.springframework.data.mongodb.repository.config.EnableMongoRepositories;
+
+import web.mediscreen.historic.repositories.HistoricRepositories;
+
+@EnableMongoRepositories(basePackageClasses = HistoricRepositories.class)
+public class MongoDbConfig {
+
+}
