+++ historic/src/main/java/web/mediscreen/historic/exception/CustomGlobalExceptionHandler.java
@@ -1,0 +1,158 @@
+package web.mediscreen.historic.exception;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Set;
+
+import javax.servlet.http.HttpServletResponse;
+
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpMethod;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.validation.FieldError;
+import org.springframework.validation.ObjectError;
+import org.springframework.web.HttpMediaTypeNotSupportedException;
+import org.springframework.web.HttpRequestMethodNotSupportedException;
+import org.springframework.web.bind.MethodArgumentNotValidException;
+import org.springframework.web.bind.MissingServletRequestParameterException;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.context.request.WebRequest;
+import org.springframework.web.method.annotation.MethodArgumentTypeMismatchException;
+import org.springframework.web.multipart.support.MissingServletRequestPartException;
+import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
+
+import feign.FeignException;
+
+/**
+ * @author Nico
+ *         <p>
+ *         Exception handler for global API exception
+ *         </p>
+ *
+ */
+@ControllerAdvice
+public class CustomGlobalExceptionHandler extends ResponseEntityExceptionHandler {
+
+	// 400
+
+	@Override
+	protected ResponseEntity<Object> handleMethodArgumentNotValid(final MethodArgumentNotValidException ex,
+			final HttpHeaders headers, final HttpStatus status, final WebRequest request) {
+		logger.info(ex.getClass().getName());
+		//
+		final List<String> errors = new ArrayList<>();
+		for (final FieldError error : ex.getBindingResult().getFieldErrors()) {
+			errors.add(error.getField() + ": " + error.getDefaultMessage());
+		}
+		for (final ObjectError error : ex.getBindingResult().getGlobalErrors()) {
+			errors.add(error.getObjectName() + ": " + error.getDefaultMessage());
+		}
+		final var apiError = new ApiError(HttpStatus.BAD_REQUEST, ex.getLocalizedMessage(), errors);
+		return handleExceptionInternal(ex, apiError, headers, apiError.getStatus(), request);
+	}
+
+	@Override
+	protected ResponseEntity<Object> handleMissingServletRequestPart(final MissingServletRequestPartException ex,
+			final HttpHeaders headers, final HttpStatus status, final WebRequest request) {
+		logger.info(ex.getClass().getName());
+		//
+		final String error = ex.getRequestPartName() + " part is missing";
+		final var apiError = new ApiError(HttpStatus.BAD_REQUEST, ex.getLocalizedMessage(), error);
+		return new ResponseEntity<>(apiError, new HttpHeaders(), apiError.getStatus());
+	}
+
+	@Override
+	protected ResponseEntity<Object> handleMissingServletRequestParameter(
+			final MissingServletRequestParameterException ex, final HttpHeaders headers, final HttpStatus status,
+			final WebRequest request) {
+		logger.info(ex.getClass().getName());
+		//
+		final String error = ex.getParameterName() + " parameter is missing";
+		final var apiError = new ApiError(HttpStatus.BAD_REQUEST, ex.getLocalizedMessage(), error);
+		return new ResponseEntity<>(apiError, new HttpHeaders(), apiError.getStatus());
+	}
+
+	//
+
+	@ExceptionHandler(FeignException.BadRequest.class)
+	public ResponseEntity<Object> handleFeignStatusException(FeignException ex, HttpServletResponse response) {
+		logger.info(ex.getClass().getName());
+		final var error = "Error while retrieving the patient";
+		final var apiError = new ApiError(HttpStatus.NOT_FOUND, ex.getLocalizedMessage(), error);
+		return new ResponseEntity<>(apiError, new HttpHeaders(), apiError.getStatus());
+	}
+
+	//
+
+	@ExceptionHandler({ MethodArgumentTypeMismatchException.class })
+	public ResponseEntity<Object> handleMethodArgumentTypeMismatch(final MethodArgumentTypeMismatchException ex,
+			final WebRequest request) {
+		logger.info(ex.getClass().getName());
+		//
+		Class<?> requiredType = ex.getRequiredType();
+		String nameType = null;
+		if (requiredType == null) {
+			nameType = "Unknown";
+		} else {
+			nameType = requiredType.getName();
+		}
+		final String error = ex.getName() + " should be of type " + nameType;
+
+		final var apiError = new ApiError(HttpStatus.BAD_REQUEST, ex.getLocalizedMessage(), error);
+		return new ResponseEntity<>(apiError, new HttpHeaders(), apiError.getStatus());
+	}
+
+	// 405
+
+	@Override
+	protected ResponseEntity<Object> handleHttpRequestMethodNotSupported(
+			final HttpRequestMethodNotSupportedException ex, final HttpHeaders headers, final HttpStatus status,
+			final WebRequest request) {
+		logger.info(ex.getClass().getName());
+		//
+		final var builder = new StringBuilder();
+		builder.append(ex.getMethod());
+		builder.append(" method is not supported for this request. Supported methods are ");
+		Set<HttpMethod> requiredHttp = ex.getSupportedHttpMethods();
+		String response = null;
+		if (requiredHttp == null) {
+			response = "Unknown";
+		} else {
+			requiredHttp.forEach(t -> builder.append(t + " "));
+			response = builder.toString();
+		}
+		final var apiError = new ApiError(HttpStatus.METHOD_NOT_ALLOWED, ex.getLocalizedMessage(), response);
+		return new ResponseEntity<>(apiError, new HttpHeaders(), apiError.getStatus());
+	}
+
+	// 415
+
+	@Override
+	protected ResponseEntity<Object> handleHttpMediaTypeNotSupported(final HttpMediaTypeNotSupportedException ex,
+			final HttpHeaders headers, final HttpStatus status, final WebRequest request) {
+		logger.info(ex.getClass().getName());
+		//
+		final var builder = new StringBuilder();
+		builder.append(ex.getContentType());
+		builder.append(" media type is not supported. Supported media types are ");
+		ex.getSupportedMediaTypes().forEach(t -> builder.append(t + " "));
+
+		final var apiError = new ApiError(HttpStatus.UNSUPPORTED_MEDIA_TYPE, ex.getLocalizedMessage(),
+				builder.substring(0, builder.length() - 2));
+		return new ResponseEntity<>(apiError, new HttpHeaders(), apiError.getStatus());
+	}
+
+	// 500
+
+	@ExceptionHandler({ Exception.class })
+	public ResponseEntity<Object> handleAll(final Exception ex, final WebRequest request) {
+		logger.info(ex.getClass().getName());
+		logger.error("error", ex);
+		//
+		final var apiError = new ApiError(HttpStatus.INTERNAL_SERVER_ERROR, ex.getLocalizedMessage(),
+				"error occurred");
+		return new ResponseEntity<>(apiError, new HttpHeaders(), apiError.getStatus());
+	}
+}
