+++ historic/src/main/java/web/mediscreen/historic/exception/HistoryNotFoundException.java
@@ -1,0 +1,30 @@
+package web.mediscreen.historic.exception;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(HttpStatus.BAD_REQUEST)
+public class HistoryNotFoundException extends Exception {
+
+    /**
+     * 
+     */
+    private static final long serialVersionUID = 1L;
+
+    private final String message;
+
+    /**
+     * @param message to return
+     */
+    public HistoryNotFoundException(String message) {
+	this.message = message;
+    }
+
+    /**
+     * @return the message
+     */
+    @Override
+    public String getMessage() {
+	return message;
+    }
+}
